requirements.txt
anyio==4.8.0
click==8.1.8
fastapi==0.95.2
h11==0.14.0
idna==3.10
pika==1.3.2
pydantic==1.10.18
python-decouple==3.5
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.22.0
setuptools==75.1.0
cryptography==44.0.2
loguru